+++ src/mrpro/algorithms/reconstruction/DirectReconstruction.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Literal, Self
+from typing import Self
 
 from mrpro.algorithms.prewhiten_kspace import prewhiten_kspace
 from mrpro.algorithms.reconstruction.Reconstruction import Reconstruction
@@ -30,21 +30,9 @@ from mrpro.operators.LinearOperator import LinearOperator
 class DirectReconstruction(Reconstruction):
     """Direct Reconstruction by Adjoint Fourier Transform."""
 
-    dcf: DcfData | None
-    """Density Compensation Data."""
-
-    csm: CsmData | None
-    """Coil Sensitivity Data."""
-
-    noise: KNoise | None
-    """Noise Data used for prewhitening."""
-
-    fourier_op: LinearOperator
-    """Fourier Operator used for the adjoint."""
-
     def __init__(
         self,
-        fourier_operator: LinearOperator,
+        fourier_op: LinearOperator,
         csm: None | CsmData = None,
         noise: None | KNoise = None,
         dcf: DcfData | None = None,
@@ -64,7 +52,7 @@ class DirectReconstruction(Reconstruction):
             Also set to None, if the FourierOperator is already density compensated.
         """
         super().__init__()
-        self.fourier_op = fourier_operator
+        self.fourier_op = fourier_op
         # TODO: Make this buffers once DataBufferMixin is merged
         self.dcf = dcf
         self.csm = csm
@@ -95,43 +83,6 @@ class DirectReconstruction(Reconstruction):
             self.recalculate_csm_walsh(kdata, noise=False)
         return self
 
-    def recalculate_fourierop(self, kdata: KData) -> Self:
-        """Update (in place) the Fourier Operator, e.g. for a new trajectory.
-
-        Also recalculates the DCF.
-
-        Parameters
-        ----------
-        kdata
-            KData to determine trajectory and recon/encoding matrix from.
-        """
-        self.fourier_op = FourierOp.from_kdata(kdata)
-        self.dcf = DcfData.from_traj_voronoi(kdata.traj)
-        return self
-
-    def recalculate_csm_walsh(self, kdata: KData, noise: KNoise | None | Literal[False] = None) -> Self:
-        """Update (in place) the CSM from KData using Walsh.
-
-        Parameters
-        ----------
-        kdata
-            KData used for adjoint reconstruction, which is then used for
-            Walsh CSM estimation.
-        noise
-            Noise measurement for prewhitening.
-            If None, self.noise (if previously set) is used.
-            If False, no prewithening is performed even if self.noise is set.
-            Use this if the kdata is already prewhitened.
-        """
-        if noise is False:
-            noise = None
-        elif noise is None:
-            noise = self.noise
-        adjoint = type(self)(self.fourier_op, dcf=self.dcf, noise=noise)
-        image = adjoint(kdata)
-        self.csm = CsmData.from_idata_walsh(image)
-        return self
-
     def forward(self, kdata: KData) -> IData:
         """Apply the reconstruction.
 
@@ -144,15 +95,4 @@ class DirectReconstruction(Reconstruction):
         -------
             the reconstruced image.
         """
-        device = kdata.data.device
-        if self.noise is not None:
-            kdata = prewhiten_kspace(kdata, self.noise.to(device))
-        operator = self.fourier_op
-        if self.csm is not None:
-            operator = operator @ self.csm.as_operator()
-        if self.dcf is not None:
-            operator = self.dcf.as_operator() @ operator
-        operator = operator.to(device)
-        (img_tensor,) = operator.H(kdata.data)
-        img = IData.from_tensor_and_kheader(img_tensor, kdata.header)
-        return img
+        return self.pseudo_inverse(kdata)
